revert symlink changes which switched to using argMap for testing args (microsoft/vscode-python#23148)
fixes https://github.com/microsoft/vscode-python/issues/22999
reverts some changes made here:
https://github.com/microsoft/vscode-python/pull/22885

+++ extensions/positron-python/src/test/testing/common/testingAdapter.test.ts
@@ -72,13 +72,17 @@ suite('End to End Tests: test adapters', () => {
         // create symlink for specific symlink test
         const target = rootPathSmallWorkspace;
         const dest = rootPathDiscoverySymlink;
-        fs.symlink(target, dest, 'dir', (err) => {
-            if (err) {
-                console.error(err);
-            } else {
-                console.log('Symlink created successfully for end to end tests.');
-            }
-        });
+        try {
+            fs.symlink(target, dest, 'dir', (err) => {
+                if (err) {
+                    console.error(err);
+                } else {
+                    console.log('Symlink created successfully for end to end tests.');
+                }
+            });
+        } catch (err) {
+            console.error(err);
+        }
     });
 
     setup(async () => {
@@ -117,13 +121,17 @@ suite('End to End Tests: test adapters', () => {
     suiteTeardown(async () => {
         // remove symlink
         const dest = rootPathDiscoverySymlink;
-        fs.unlink(dest, (err) => {
-            if (err) {
-                console.error(err);
-            } else {
-                console.log('Symlink removed successfully after tests.');
-            }
-        });
+        if (fs.existsSync(dest)) {
+            fs.unlink(dest, (err) => {
+                if (err) {
+                    console.error(err);
+                } else {
+                    console.log('Symlink removed successfully after tests.');
+                }
+            });
+        } else {
+            console.log('Symlink was not found to remove after tests, exiting successfully');
+        }
     });
     test('unittest discovery adapter small workspace', async () => {
         // result resolver and saved data for assertions
@@ -293,6 +301,7 @@ suite('End to End Tests: test adapters', () => {
             resultResolver,
             envVarsService,
         );
+        configService.getSettings(workspaceUri).testing.pytestArgs = [];
 
         await discoveryAdapter.discoverTests(workspaceUri, pythonExecFactory).finally(() => {
             // verification after discovery is complete
@@ -372,6 +381,7 @@ suite('End to End Tests: test adapters', () => {
 
         // set workspace to test workspace folder
         workspaceUri = Uri.parse(rootPathLargeWorkspace);
+        configService.getSettings(workspaceUri).testing.pytestArgs = [];
 
         await discoveryAdapter.discoverTests(workspaceUri, pythonExecFactory).finally(() => {
             // verification after discovery is complete
@@ -558,6 +568,7 @@ suite('End to End Tests: test adapters', () => {
         };
         // set workspace to test workspace folder
         workspaceUri = Uri.parse(rootPathSmallWorkspace);
+        configService.getSettings(workspaceUri).testing.pytestArgs = [];
 
         // run pytest execution
         const executionAdapter = new PytestTestExecutionAdapter(
@@ -648,6 +659,7 @@ suite('End to End Tests: test adapters', () => {
 
         // set workspace to test workspace folder
         workspaceUri = Uri.parse(rootPathLargeWorkspace);
+        configService.getSettings(workspaceUri).testing.pytestArgs = [];
 
         // generate list of test_ids
         const testIds: string[] = [];
@@ -728,6 +740,7 @@ suite('End to End Tests: test adapters', () => {
 
         // set workspace to test workspace folder
         workspaceUri = Uri.parse(rootPathDiscoveryErrorWorkspace);
+        configService.getSettings(workspaceUri).testing.unittestArgs = ['-s', '.', '-p', '*test*.py'];
 
         const discoveryAdapter = new UnittestTestDiscoveryAdapter(
             pythonTestServer,
@@ -799,6 +812,8 @@ suite('End to End Tests: test adapters', () => {
 
         // set workspace to test workspace folder
         workspaceUri = Uri.parse(rootPathDiscoveryErrorWorkspace);
+        configService.getSettings(workspaceUri).testing.pytestArgs = [];
+
         await discoveryAdapter.discoverTests(workspaceUri, pythonExecFactory).finally(() => {
             // verification after discovery is complete
             assert.ok(
@@ -860,6 +875,7 @@ suite('End to End Tests: test adapters', () => {
 
         // set workspace to test workspace folder
         workspaceUri = Uri.parse(rootPathErrorWorkspace);
+        configService.getSettings(workspaceUri).testing.unittestArgs = ['-s', '.', '-p', '*test*.py'];
 
         // run pytest execution
         const executionAdapter = new UnittestTestExecutionAdapter(
@@ -921,6 +937,7 @@ suite('End to End Tests: test adapters', () => {
 
         // set workspace to test workspace folder
         workspaceUri = Uri.parse(rootPathErrorWorkspace);
+        configService.getSettings(workspaceUri).testing.pytestArgs = [];
 
         // run pytest execution
         const executionAdapter = new PytestTestExecutionAdapter(
